refactor-process-context
